Fix a memory leak bug in demo for ViewBadger. And add leakcanary in project.
