+++ app/build.gradle
@@ -32,4 +32,9 @@ dependencies {
     testCompile 'junit:junit:4.12'
     compile project(':widget')
     compile files('libs/android-viewbadger.jar')
+
+    // leakcanary
+    debugCompile 'com.squareup.leakcanary:leakcanary-android:1.5'
+    releaseCompile 'com.squareup.leakcanary:leakcanary-android-no-op:1.5'
+    testCompile 'com.squareup.leakcanary:leakcanary-android-no-op:1.5'
 }
