+++ app/src/main/res/layout/activity_main.xml
@@ -11,7 +11,7 @@
         android:layout_marginRight="@dimen/activity_horizontal_margin"
         android:layout_marginTop="@dimen/activity_vertical_margin"
         android:orientation="vertical"
-        tools:context=".MainActivity">
+        tools:context=".features.MainActivity">
 
         <Button
             android:id="@+id/btn_style"
