+++ app/src/main/res/layout/activity_style.xml
@@ -12,7 +12,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="vertical"
-            tools:context=".style.StyleActivity">
+            tools:context=".features.style.StyleActivity">
 
             <TextView
                 style="@style/text_black"
